the unholy rewrite pt.3 - CSS
removed a bunch of code that I don’t care for.
renamed most of the variables to a more organized fashion - really
liking this.
removed line mixin - no need for it
added font-size mixin
added p{margin-top:0;} to normalize
—
renamed the classes - back to the old style of elements inside the
major parent
using is-parent and is-expanded instead of the long form
moved everything inside the small screen breakpoint - wont stay like
this, but for now it works
reorganized all the small screen stuff to be together

HAVE NOT STARTED 768px + styles yet