Rewrite
This is a full rewrite from the ground up of a accessible mega menu
style navigation system. It is also responsive, in that it condenses
down into a left menu.
Based on adobes accessible mega menu, but removed dependencies with js,
removed all css and build from the ground up. What is still remnant of
the adobe code is the general html markup for the menus themselves.